requirements.txt
# Scraping
beautifulsoup4==4.12.2 
requests==2.32.3

# EDA & Data Preparation
pandas==2.0.3
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.1
wordcloud==1.9.3

# RAG
langchain==0.2.2
langchain-qdrant==0.1.3
langchain-huggingface==0.0.3
langchain-ollama==0.1.1
langchain-core==0.2.43
qdrant-client==1.12.1

# Others
python-dotenv==1.0.1 
unstructured==0.14.4 

# install markdown depenendies with: `pip install "unstructured[md]"` after install the requirements file. Leave this line commented out. 